refactor: externalize component templates and styles
