+++ src/app/app.component.html
@@ -1,0 +1,56 @@
+<div class="app">
+      <div class="main-content" [class.with-nav]="showNavigation">
+        <router-outlet></router-outlet>
+      </div>
+      
+      <nav class="bottom-nav" *ngIf="showNavigation">
+        <button 
+          class="nav-item" 
+          [class.active]="currentRoute === '/dashboard'"
+          (click)="navigate('/dashboard')"
+        >
+          <span class="nav-icon">🏠</span>
+          <span class="nav-label">Home</span>
+        </button>
+        <button 
+          class="nav-item" 
+          [class.active]="currentRoute === '/plan'"
+          (click)="navigate('/plan')"
+        >
+          <span class="nav-icon">📋</span>
+          <span class="nav-label">Plan</span>
+        </button>
+        <button 
+          class="nav-item" 
+          [class.active]="currentRoute === '/log'"
+          (click)="navigate('/log')"
+        >
+          <span class="nav-icon">📝</span>
+          <span class="nav-label">Log</span>
+        </button>
+        <button 
+          class="nav-item" 
+          [class.active]="currentRoute === '/history'"
+          (click)="navigate('/history')"
+        >
+          <span class="nav-icon">📊</span>
+          <span class="nav-label">History</span>
+        </button>
+        <button 
+          class="nav-item" 
+          [class.active]="currentRoute === '/progress'"
+          (click)="navigate('/progress')"
+        >
+          <span class="nav-icon">📈</span>
+          <span class="nav-label">Progress</span>
+        </button>
+        <button 
+          class="nav-item" 
+          [class.active]="currentRoute === '/profile'"
+          (click)="navigate('/profile')"
+        >
+          <span class="nav-icon">👤</span>
+          <span class="nav-label">Profile</span>
+        </button>
+      </nav>
+    </div>
