+++ src/app/app.component.scss
@@ -1,0 +1,92 @@
+.app {
+      min-height: 100vh;
+      display: flex;
+      flex-direction: column;
+    }
+
+    .main-content {
+      flex: 1;
+      background: #f3f4f6;
+      min-height: 100vh;
+    }
+
+    .main-content.with-nav {
+      padding-bottom: 80px;
+    }
+
+    .bottom-nav {
+      position: fixed;
+      bottom: 0;
+      left: 0;
+      right: 0;
+      background: white;
+      border-top: 1px solid #e5e7eb;
+      display: flex;
+      justify-content: space-around;
+      padding: 0.5rem 0.25rem;
+      box-shadow: 0 -4px 6px -1px rgba(0, 0, 0, 0.1);
+      z-index: 1000;
+    }
+
+    .nav-item {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      padding: 0.5rem;
+      border: none;
+      background: transparent;
+      cursor: pointer;
+      transition: all 0.2s;
+      border-radius: 0.5rem;
+      flex: 1;
+      max-width: 80px;
+    }
+
+    .nav-item:hover {
+      background: #f3f4f6;
+    }
+
+    .nav-item.active {
+      background: #eff6ff;
+      color: #3b82f6;
+    }
+
+    .nav-icon {
+      font-size: 1.25rem;
+      margin-bottom: 0.25rem;
+    }
+
+    .nav-label {
+      font-size: 0.75rem;
+      font-weight: 500;
+      color: inherit;
+    }
+
+    .nav-item.active .nav-label {
+      color: #3b82f6;
+    }
+
+    @media (min-width: 1024px) {
+      .bottom-nav {
+        display: none;
+      }
+
+      .main-content.with-nav {
+        padding-bottom: 0;
+      }
+    }
+
+    @media (max-width: 480px) {
+      .nav-label {
+        font-size: 0.625rem;
+      }
+
+      .nav-icon {
+        font-size: 1rem;
+      }
+
+      .nav-item {
+        padding: 0.375rem;
+      }
+    }
