+++ src/app/components/log-workout/log-workout.component.scss
@@ -1,0 +1,294 @@
+.log-workout {
+      padding: 1rem;
+      max-width: 1200px;
+      margin: 0 auto;
+    }
+
+    .page-header {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      margin-bottom: 2rem;
+    }
+
+    .page-header h1 {
+      color: #1f2937;
+      font-size: 1.875rem;
+      font-weight: 700;
+      margin: 0;
+    }
+
+    .workout-timer {
+      display: flex;
+      align-items: center;
+      gap: 0.5rem;
+      background: #f3f4f6;
+      padding: 0.5rem 1rem;
+      border-radius: 0.75rem;
+    }
+
+    .timer-icon {
+      font-size: 1.25rem;
+    }
+
+    .time {
+      font-weight: 600;
+      color: #1f2937;
+      font-size: 1.125rem;
+    }
+
+    .day-selection {
+      background: white;
+      padding: 2rem;
+      border-radius: 1rem;
+      box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+    }
+
+    .day-selection h2 {
+      color: #1f2937;
+      margin-bottom: 1.5rem;
+    }
+
+    .day-cards {
+      display: grid;
+      grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+      gap: 1rem;
+    }
+
+    .day-card {
+      background: #f9fafb;
+      padding: 1.5rem;
+      border-radius: 0.75rem;
+      border: 2px solid #e5e7eb;
+      cursor: pointer;
+      transition: all 0.2s;
+      text-align: center;
+    }
+
+    .day-card:hover {
+      border-color: #3b82f6;
+      background: #eff6ff;
+    }
+
+    .day-card h3 {
+      color: #1f2937;
+      margin-bottom: 0.5rem;
+    }
+
+    .day-badge {
+      background: #3b82f6;
+      color: white;
+      padding: 0.25rem 0.75rem;
+      border-radius: 9999px;
+      font-size: 0.875rem;
+      display: inline-block;
+      margin-bottom: 0.5rem;
+    }
+
+    .exercise-count {
+      color: #6b7280;
+      font-size: 0.875rem;
+      margin: 0;
+    }
+
+    .workout-logging {
+      background: white;
+      padding: 2rem;
+      border-radius: 1rem;
+      box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+    }
+
+    .workout-header {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      margin-bottom: 2rem;
+      flex-wrap: wrap;
+      gap: 1rem;
+    }
+
+    .workout-header h2 {
+      color: #1f2937;
+      margin: 0;
+    }
+
+    .workout-date {
+      color: #6b7280;
+      margin: 0;
+    }
+
+    .change-day-btn {
+      padding: 0.5rem 1rem;
+      background: #6b7280;
+      color: white;
+      border: none;
+      border-radius: 0.375rem;
+      cursor: pointer;
+      font-size: 0.875rem;
+    }
+
+    .exercise-card {
+      background: #f9fafb;
+      padding: 1.5rem;
+      border-radius: 0.75rem;
+      margin-bottom: 1.5rem;
+      border: 1px solid #e5e7eb;
+    }
+
+    .exercise-header {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      margin-bottom: 1rem;
+    }
+
+    .exercise-header h3 {
+      color: #1f2937;
+      margin: 0;
+    }
+
+    .exercise-status {
+      background: #3b82f6;
+      color: white;
+      padding: 0.25rem 0.75rem;
+      border-radius: 9999px;
+      font-size: 0.875rem;
+      font-weight: 600;
+    }
+
+    .sets-table {
+      background: white;
+      border-radius: 0.5rem;
+      overflow: hidden;
+      border: 1px solid #e5e7eb;
+    }
+
+    .table-header {
+      display: grid;
+      grid-template-columns: 0.5fr 1fr 1fr 1fr 0.5fr;
+      gap: 0.5rem;
+      padding: 0.75rem;
+      background: #f3f4f6;
+      font-weight: 600;
+      color: #374151;
+      font-size: 0.875rem;
+    }
+
+    .set-row {
+      display: grid;
+      grid-template-columns: 0.5fr 1fr 1fr 1fr 0.5fr;
+      gap: 0.5rem;
+      padding: 0.75rem;
+      border-top: 1px solid #e5e7eb;
+      align-items: center;
+    }
+
+    .set-number {
+      font-weight: 600;
+      color: #6b7280;
+    }
+
+    .previous-data {
+      color: #6b7280;
+      font-size: 0.875rem;
+    }
+
+    .reps-input, .weight-input {
+      padding: 0.5rem;
+      border: 1px solid #d1d5db;
+      border-radius: 0.375rem;
+      text-align: center;
+      width: 100%;
+      box-sizing: border-box;
+    }
+
+    .completed-checkbox {
+      width: 1.25rem;
+      height: 1.25rem;
+      justify-self: center;
+    }
+
+    .workout-notes {
+      margin: 2rem 0;
+    }
+
+    .workout-notes label {
+      display: block;
+      margin-bottom: 0.5rem;
+      font-weight: 500;
+      color: #374151;
+    }
+
+    .workout-notes textarea {
+      width: 100%;
+      padding: 0.75rem;
+      border: 1px solid #d1d5db;
+      border-radius: 0.5rem;
+      resize: vertical;
+      box-sizing: border-box;
+    }
+
+    .workout-actions {
+      display: flex;
+      gap: 1rem;
+      justify-content: flex-end;
+    }
+
+    .cancel-btn {
+      padding: 0.75rem 1.5rem;
+      background: #6b7280;
+      color: white;
+      border: none;
+      border-radius: 0.5rem;
+      cursor: pointer;
+    }
+
+    .save-btn {
+      padding: 0.75rem 1.5rem;
+      background: #10b981;
+      color: white;
+      border: none;
+      border-radius: 0.5rem;
+      font-weight: 600;
+      cursor: pointer;
+      transition: background-color 0.2s;
+    }
+
+    .save-btn:hover:not(:disabled) {
+      background: #059669;
+    }
+
+    .save-btn:disabled {
+      opacity: 0.6;
+      cursor: not-allowed;
+    }
+
+    @media (max-width: 768px) {
+      .log-workout {
+        padding: 1rem 0.5rem;
+      }
+
+      .page-header {
+        flex-direction: column;
+        gap: 1rem;
+      }
+
+      .workout-header {
+        flex-direction: column;
+        align-items: flex-start;
+      }
+
+      .day-cards {
+        grid-template-columns: 1fr;
+      }
+
+      .table-header, .set-row {
+        grid-template-columns: 0.4fr 0.8fr 0.8fr 1fr 0.4fr;
+        gap: 0.25rem;
+        font-size: 0.75rem;
+      }
+
+      .workout-actions {
+        flex-direction: column;
+      }
+    }
