+++ src/app/components/login/login.component.html
@@ -1,0 +1,72 @@
+<div class="login-container">
+      <div class="login-card">
+        <h1>Workout Tracker</h1>
+        <div class="tabs">
+          <button 
+            class="tab-button" 
+            [class.active]="isLoginMode" 
+            (click)="isLoginMode = true"
+          >
+            Login
+          </button>
+          <button 
+            class="tab-button" 
+            [class.active]="!isLoginMode" 
+            (click)="isLoginMode = false"
+          >
+            Sign Up
+          </button>
+        </div>
+
+        <form (ngSubmit)="onSubmit()" #form="ngForm">
+          <div class="form-group" *ngIf="!isLoginMode">
+            <label for="name">Name</label>
+            <input 
+              type="text" 
+              id="name" 
+              name="name"
+              [(ngModel)]="formData.name"
+              required
+              placeholder="Enter your name"
+            />
+          </div>
+
+          <div class="form-group">
+            <label for="email">Email</label>
+            <input 
+              type="email" 
+              id="email" 
+              name="email"
+              [(ngModel)]="formData.email"
+              required
+              placeholder="Enter your email"
+            />
+          </div>
+
+          <div class="form-group">
+            <label for="password">Password</label>
+            <input 
+              type="password" 
+              id="password" 
+              name="password"
+              [(ngModel)]="formData.password"
+              required
+              placeholder="Enter your password"
+            />
+          </div>
+
+          <button 
+            type="submit" 
+            class="submit-button"
+            [disabled]="!form.form.valid || isLoading"
+          >
+            <span *ngIf="isLoading">Loading...</span>
+            <span *ngIf="!isLoading">{{ isLoginMode ? 'Login' : 'Sign Up' }}</span>
+          </button>
+        </form>
+
+        <div class="demo-note">
+          <p><strong>Demo Mode:</strong> Use any email/password to sign in</p>
+        </div>
+      </div>
+    </div>
