+++ src/app/components/profile/profile.component.scss
@@ -1,119 +1,4 @@
-import { Component, inject } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { Router } from '@angular/router';
-import { AuthService } from '../services/auth.service';
-
-@Component({
-  selector: 'app-profile',
-  standalone: true,
-  imports: [CommonModule],
-  template: `
-    <div class="profile">
-      <header class="profile-header">
-        <div class="user-avatar">{{ getUserInitials() }}</div>
-        <div class="user-info">
-          <h1>{{ currentUser?.name }}</h1>
-          <p>{{ currentUser?.email }}</p>
-          <p class="member-since">Member since {{ formatDate(currentUser?.createdAt) }}</p>
-        </div>
-      </header>
-
-      <div class="profile-sections">
-        <div class="section">
-          <h2>Account Settings</h2>
-          <div class="setting-item">
-            <span class="setting-label">Email Notifications</span>
-            <label class="switch">
-              <input type="checkbox" checked>
-              <span class="slider"></span>
-            </label>
-          </div>
-          <div class="setting-item">
-            <span class="setting-label">Workout Reminders</span>
-            <label class="switch">
-              <input type="checkbox" checked>
-              <span class="slider"></span>
-            </label>
-          </div>
-          <div class="setting-item">
-            <span class="setting-label">Progress Reports</span>
-            <label class="switch">
-              <input type="checkbox">
-              <span class="slider"></span>
-            </label>
-          </div>
-        </div>
-
-        <div class="section">
-          <h2>Workout Preferences</h2>
-          <div class="preference-item">
-            <label>Default Rest Time (seconds)</label>
-            <select>
-              <option value="60">60</option>
-              <option value="90">90</option>
-              <option value="120" selected>120</option>
-              <option value="180">180</option>
-            </select>
-          </div>
-          <div class="preference-item">
-            <label>Weight Unit</label>
-            <select>
-              <option value="kg" selected>Kilograms (kg)</option>
-              <option value="lbs">Pounds (lbs)</option>
-            </select>
-          </div>
-          <div class="preference-item">
-            <label>Week Start Day</label>
-            <select>
-              <option value="sunday">Sunday</option>
-              <option value="monday" selected>Monday</option>
-            </select>
-          </div>
-        </div>
-
-        <div class="section">
-          <h2>Data & Privacy</h2>
-          <div class="action-item">
-            <div>
-              <span class="action-title">Export Workout Data</span>
-              <p class="action-description">Download all your workout data as a CSV file</p>
-            </div>
-            <button class="secondary-btn">Export</button>
-          </div>
-          <div class="action-item">
-            <div>
-              <span class="action-title">Clear Workout History</span>
-              <p class="action-description">This will permanently delete all your workout logs</p>
-            </div>
-            <button class="danger-btn">Clear</button>
-          </div>
-        </div>
-
-        <div class="section">
-          <h2>About</h2>
-          <div class="about-content">
-            <p><strong>Workout Tracker</strong></p>
-            <p>Version 1.0.0</p>
-            <p>Built with Angular 20 and modern web technologies</p>
-            <div class="links">
-              <a href="#" class="link">Privacy Policy</a>
-              <a href="#" class="link">Terms of Service</a>
-              <a href="#" class="link">Support</a>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div class="logout-section">
-        <button class="logout-btn" (click)="logout()">
-          <span class="logout-icon">🚪</span>
-          Sign Out
-        </button>
-      </div>
-    </div>
-  `,
-  styles: [`
-    .profile {
+.profile {
       padding: 1rem;
       max-width: 800px;
       margin: 0 auto;
@@ -397,37 +282,3 @@ import { AuthService } from '../services/auth.service';
         gap: 0.5rem;
       }
     }
-  `]
-})
-export class ProfileComponent {
-    private authService: AuthService = inject(AuthService)
-
-  currentUser = this.authService.currentUserValue;
-
-  constructor(
-    private router: Router
-  ) {}
-
-  getUserInitials(): string {
-    if (!this.currentUser?.name) return 'U';
-    return this.currentUser.name
-      .split(' ')
-      .map(name => name[0])
-      .join('')
-      .toUpperCase()
-      .slice(0, 2);
-  }
-
-  formatDate(date: Date | undefined): string {
-    if (!date) return '';
-    return new Date(date).toLocaleDateString('en-US', {
-      year: 'numeric',
-      month: 'long'
-    });
-  }
-
-  logout(): void {
-    this.authService.logout();
-    this.router.navigate(['/login']);
-  }
-}
