+++ src/app/components/workout-history/workout-history.component.html
@@ -1,0 +1,82 @@
+<div class="workout-history">
+      <header class="page-header">
+        <h1>Workout History</h1>
+        <div class="filter-controls">
+          <select [(ngModel)]="selectedFilter" (change)="applyFilter()">
+            <option value="all">All Time</option>
+            <option value="week">This Week</option>
+            <option value="month">This Month</option>
+            <option value="3months">Last 3 Months</option>
+          </select>
+        </div>
+      </header>
+
+      <div class="stats-overview">
+        <div class="stat-card">
+          <div class="stat-value">{{ filteredLogs.length }}</div>
+          <div class="stat-label">Total Workouts</div>
+        </div>
+        <div class="stat-card">
+          <div class="stat-value">{{ getTotalDuration() }}h</div>
+          <div class="stat-label">Time Spent</div>
+        </div>
+        <div class="stat-card">
+          <div class="stat-value">{{ getUniqueExercises() }}</div>
+          <div class="stat-label">Unique Exercises</div>
+        </div>
+        <div class="stat-card">
+          <div class="stat-value">{{ getAverageWorkout() }}min</div>
+          <div class="stat-label">Avg Duration</div>
+        </div>
+      </div>
+
+      <div class="workout-list" *ngIf="filteredLogs.length > 0">
+        <div class="workout-item" *ngFor="let log of filteredLogs">
+          <div class="workout-header" (click)="toggleExpanded(log.id)">
+            <div class="workout-info">
+              <h3>{{ log.dayName }}</h3>
+              <p class="workout-date">{{ log.date | date:'MMM d, y - EEEE' }}</p>
+            </div>
+            <div class="workout-summary">
+              <span class="duration">{{ log.duration }}min</span>
+              <span class="exercise-count">{{ log.exercises.length }} exercises</span>
+              <span class="expand-icon" [class.rotated]="expandedWorkouts.has(log.id)">▼</span>
+            </div>
+          </div>
+
+          <div class="workout-details" *ngIf="expandedWorkouts.has(log.id)">
+            <div class="exercise-details" *ngFor="let exercise of log.exercises">
+              <h4>{{ exercise.name }}</h4>
+              <div class="sets-summary">
+                <div class="set-item" *ngFor="let set of exercise.sets">
+                  <span class="set-label">Set {{ set.setNumber }}:</span>
+                  <span class="set-data">{{ set.reps }} reps X {{ set.weight }}kg</span>
+                  <span class="completed-badge" *ngIf="set.completed">✓</span>
+                </div>
+              </div>
+            </div>
+            <div class="workout-notes" *ngIf="log.notes">
+              <h4>Notes:</h4>
+              <p>{{ log.notes }}</p>
+            </div>
+          </div>
+        </div>
+      </div>
+
+      <div class="empty-state" *ngIf="filteredLogs.length === 0">
+        <h2>No Workouts Found</h2>
+        <p *ngIf="selectedFilter === 'all'">You haven't logged any workouts yet.</p>
+        <p *ngIf="selectedFilter !== 'all'">No workouts found for the selected time period.</p>
+        <button class="start-workout-btn" (click)="navigateToLog()">Log Your First Workout</button>
+      </div>
+
+      <div class="workout-calendar" *ngIf="workoutLogs.length > 0">
+        <h2>Workout Calendar</h2>
+        <div class="calendar-grid">
+          <div class="calendar-day" *ngFor="let day of getCalendarDays()">
+            <span class="day-number">{{ day.date.getDate() }}</span>
+            <div class="workout-indicator" *ngIf="day.hasWorkout" [title]="day.workoutName"></div>
+          </div>
+        </div>
+      </div>
+    </div>
