+++ src/app/components/workout-history/workout-history.component.scss
@@ -1,0 +1,291 @@
+.workout-history {
+      padding: 1rem;
+      max-width: 1200px;
+      margin: 0 auto;
+    }
+
+    .page-header {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      margin-bottom: 2rem;
+    }
+
+    .page-header h1 {
+      color: #1f2937;
+      font-size: 1.875rem;
+      font-weight: 700;
+      margin: 0;
+    }
+
+    .filter-controls select {
+      padding: 0.5rem 1rem;
+      border: 2px solid #d1d5db;
+      border-radius: 0.5rem;
+      background: white;
+      color: #1f2937;
+      font-size: 0.875rem;
+    }
+
+    .stats-overview {
+      display: grid;
+      grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+      gap: 1rem;
+      margin-bottom: 2rem;
+    }
+
+    .stat-card {
+      background: white;
+      padding: 1.5rem;
+      border-radius: 1rem;
+      box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+      text-align: center;
+    }
+
+    .stat-value {
+      font-size: 2rem;
+      font-weight: 700;
+      color: #3b82f6;
+      margin-bottom: 0.5rem;
+    }
+
+    .stat-label {
+      color: #6b7280;
+      font-size: 0.875rem;
+    }
+
+    .workout-list {
+      background: white;
+      border-radius: 1rem;
+      box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+      overflow: hidden;
+      margin-bottom: 2rem;
+    }
+
+    .workout-item {
+      border-bottom: 1px solid #f3f4f6;
+    }
+
+    .workout-item:last-child {
+      border-bottom: none;
+    }
+
+    .workout-header {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      padding: 1.5rem;
+      cursor: pointer;
+      transition: background-color 0.2s;
+    }
+
+    .workout-header:hover {
+      background: #f9fafb;
+    }
+
+    .workout-info h3 {
+      color: #1f2937;
+      margin: 0 0 0.25rem 0;
+      font-size: 1.125rem;
+    }
+
+    .workout-date {
+      color: #6b7280;
+      margin: 0;
+      font-size: 0.875rem;
+    }
+
+    .workout-summary {
+      display: flex;
+      align-items: center;
+      gap: 1rem;
+      color: #6b7280;
+      font-size: 0.875rem;
+    }
+
+    .expand-icon {
+      transition: transform 0.2s;
+      font-size: 0.75rem;
+    }
+
+    .expand-icon.rotated {
+      transform: rotate(180deg);
+    }
+
+    .workout-details {
+      padding: 0 1.5rem 1.5rem 1.5rem;
+      background: #f9fafb;
+      border-top: 1px solid #f3f4f6;
+    }
+
+    .exercise-details {
+      margin-bottom: 1.5rem;
+    }
+
+    .exercise-details h4 {
+      color: #1f2937;
+      margin: 0 0 0.75rem 0;
+      font-size: 1rem;
+    }
+
+    .sets-summary {
+      display: grid;
+      gap: 0.5rem;
+    }
+
+    .set-item {
+      display: flex;
+      align-items: center;
+      gap: 0.5rem;
+      padding: 0.5rem;
+      background: white;
+      border-radius: 0.375rem;
+      font-size: 0.875rem;
+    }
+
+    .set-label {
+      font-weight: 500;
+      color: #6b7280;
+      min-width: 60px;
+    }
+
+    .set-data {
+      color: #1f2937;
+    }
+
+    .completed-badge {
+      margin-left: auto;
+      color: #10b981;
+      font-weight: 600;
+    }
+
+    .workout-notes {
+      margin-top: 1rem;
+      padding: 1rem;
+      background: #fffbeb;
+      border: 1px solid #fbbf24;
+      border-radius: 0.5rem;
+    }
+
+    .workout-notes h4 {
+      color: #92400e;
+      margin: 0 0 0.5rem 0;
+      font-size: 0.875rem;
+      font-weight: 600;
+    }
+
+    .workout-notes p {
+      color: #92400e;
+      margin: 0;
+      font-size: 0.875rem;
+    }
+
+    .empty-state {
+      text-align: center;
+      padding: 4rem 2rem;
+      background: white;
+      border-radius: 1rem;
+      box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+    }
+
+    .empty-state h2 {
+      color: #1f2937;
+      margin-bottom: 1rem;
+    }
+
+    .empty-state p {
+      color: #6b7280;
+      margin-bottom: 2rem;
+    }
+
+    .start-workout-btn {
+      padding: 0.75rem 1.5rem;
+      background: #3b82f6;
+      color: white;
+      border: none;
+      border-radius: 0.5rem;
+      font-weight: 600;
+      cursor: pointer;
+      transition: background-color 0.2s;
+    }
+
+    .start-workout-btn:hover {
+      background: #2563eb;
+    }
+
+    .workout-calendar {
+      background: white;
+      padding: 1.5rem;
+      border-radius: 1rem;
+      box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+    }
+
+    .workout-calendar h2 {
+      color: #1f2937;
+      margin-bottom: 1.5rem;
+      text-align: center;
+    }
+
+    .calendar-grid {
+      display: grid;
+      grid-template-columns: repeat(7, 1fr);
+      gap: 0.5rem;
+      max-width: 400px;
+      margin: 0 auto;
+    }
+
+    .calendar-day {
+      aspect-ratio: 1;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      background: #f9fafb;
+      border-radius: 0.375rem;
+      position: relative;
+      font-size: 0.875rem;
+    }
+
+    .day-number {
+      color: #6b7280;
+      font-weight: 500;
+    }
+
+    .workout-indicator {
+      width: 6px;
+      height: 6px;
+      background: #10b981;
+      border-radius: 50%;
+      margin-top: 2px;
+    }
+
+    @media (max-width: 768px) {
+      .workout-history {
+        padding: 1rem 0.5rem;
+      }
+
+      .page-header {
+        flex-direction: column;
+        gap: 1rem;
+        align-items: stretch;
+      }
+
+      .stats-overview {
+        grid-template-columns: repeat(2, 1fr);
+      }
+
+      .workout-header {
+        flex-direction: column;
+        align-items: flex-start;
+        gap: 0.5rem;
+      }
+
+      .workout-summary {
+        align-self: stretch;
+        justify-content: space-between;
+      }
+
+      .sets-summary {
+        grid-template-columns: 1fr;
+      }
+    }
